SQL Data Exploration – COVID case study
The objective of this case study is to explore data and extract useful information from the COVID dataset. Further we also visualize some information on Tableau. The dataset here consists of all the countries in the world with COVID cases and deaths numbers along with other information such as date, population and continent.