Add files via upload
The agenda of this project is to create a dashboard in MS Excel to display bike sales analysis. The dataset contains data of 1000 people with key parameters such as Marital status, Gender, Income, Education, Occupation, Commute distance, Region, Age and if they purchased bike or not. Below is a snapshot of the dataset. The main focus is to analyze and visualize key metrics in the dashboard to help the stakeholders make informed business decisions.